refactor(browser): separate connection and tool ownership
